style(logger): use pretty logger in development & json in production
Make sure you set NODE_ENV=development in the .env file but never set it on production
