+++ .env-example
@@ -11,3 +11,6 @@ BUNDLER_DATASOURCES_JSON= '{"mongoUrl": "", "redisUrl": ""}'
 BUNDLER_QUEUE_URL=""
 BUNDLER_IS_TRUSTWALLET_SETUP=true
 BUNDLER_FALLBACK_PROVIDER_JSON='{"137": ["", ""], "42161": ["", ""], "56": ["", ""], "10": ["", ""], "43114": ["", ""], "8453": ["", ""], "80001": ["", ""]}'
+
+// make sure this is never development when deploying to production, and delete this line
+NODE_ENV=
