+++ src/common/logger/index.ts
@@ -1,8 +1,21 @@
 import pino from "pino";
 
-const destination = pino.destination({ sync: false });
+// Use sync mode in development to avoid out-of-order logs
+const destination = pino.destination({
+  sync: process.env.NODE_ENV === "development",
+});
 
-// Create the logger using the asynchronous destination
-const logger = pino({}, destination);
+// Use the pretty logger in development, and JSON logger in production
+const logger =
+  process.env.NODE_ENV === "development"
+    ? pino({
+        transport: {
+          target: "pino-pretty",
+          options: {
+            colorize: true,
+          },
+        },
+      })
+    : pino({}, destination);
 
 export { logger };
